Added login. Added salted hashes

+++ server/routes/api/users.js
@@ -4,14 +4,17 @@ const express = require("express");
 const router = express.Router();
 const db = require("../../db");
 const authSession_1 = require("../../middleware/authSession");
+const salter_1 = require("../../salter");
 // Gets all members
-router.get("/", async function (req, res) {
+router.get("/", authSession_1.authSession, async function (req, res) {
     try {
-        db.User.find(function (err, users) {
-            if (err)
-                return console.error(err);
-            res.json(users);
-        });
+        if (!res.headersSent) {
+            db.User.find(function (err, users) {
+                if (err)
+                    return console.error(err);
+                res.json(users);
+            });
+        }
     }
     catch (err) {
         console.error(err);
@@ -46,26 +49,29 @@ router.get("/:username", authSession_1.authSession, async function (req, res) {
 router.post("/", async function (req, res) {
     try {
         let dateNow = Date.now();
+        if (!req.body.username ||
+            !req.body.password ||
+            !req.body.displayName ||
+            !req.body.biography) {
+            return res.status(400).send("Missing information");
+        }
+        let hash = salter_1.saltPassword(req.body.password);
         let newUser = new db.User({
             username: req.body.username,
-            password: req.body.password,
+            hash: hash.hash,
+            salt: hash.salt,
             displayName: req.body.displayName,
             biography: req.body.biography,
             creationDate: dateNow,
             sessionKeys: [],
         });
-        if (!newUser.username ||
-            !newUser.password ||
-            !newUser.displayName ||
-            !newUser.biography) {
-            return res.status(400).send("Missing information");
-        }
         await newUser.save((err, newUser) => {
             if (err) {
                 console.error(err);
                 return res.status(500).end();
             }
         });
+        //res.json(newUser.sendableUser());
         res.json(newUser);
     }
     catch (err) {
@@ -74,20 +80,32 @@ router.post("/", async function (req, res) {
     }
 });
 // Update member
-router.put("/:username", async function (req, res) {
+router.put("/:username", authSession_1.authSession, async function (req, res) {
     try {
-        // Get requested post
-        // Update post if there is a change
-        const userUpdate = {};
-        if (req.body.username)
-            userUpdate.username = req.body.username;
-        if (req.body.password)
-            userUpdate.password = req.body.password;
-        if (req.body.displayName)
-            userUpdate.displayName = req.body.displayName;
-        if (req.body.biography)
-            userUpdate.biography = req.body.biography;
-        db.User.updateOne({ username: req.params.username }, userUpdate);
+        if (!res.headersSent) {
+            // Get requested post
+            // Update post if there is a change
+            const user = await db.User.findOne({
+                username: req.params.username,
+            });
+            if (user != null && user != undefined) {
+                if (authSession_1.authUser(req, res, user.id)) {
+                    const userUpdate = {};
+                    if (req.body.username)
+                        user.username = req.body.username;
+                    if (req.body.displayName)
+                        user.displayName = req.body.displayName;
+                    if (req.body.biography)
+                        user.biography = req.body.biography;
+                    user.save();
+                    res.json(user.sendableUser());
+                }
+                else
+                    res.status(401).send("Not authorized to access this user");
+            }
+            else
+                res.status(404).end();
+        }
     }
     catch (err) {
         console.error(err);
@@ -95,10 +113,12 @@ router.put("/:username", async function (req, res) {
     }
 });
 //Delete member
-router.delete("/:username", (req, res) => {
+router.delete("/:username", authSession_1.authSession, (req, res) => {
     try {
-        db.User.deleteOne({ username: req.params.username });
-        res.end();
+        if (!res.headersSent) {
+            db.User.deleteOne({ username: req.params.username });
+            res.end();
+        }
     }
     catch (err) {
         console.error(err);
