+++ server/routes/api/users.ts
@@ -3,14 +3,17 @@ const router = express.Router();
 import db = require("../../db");
 import uuid = require("uuid");
 import { authSession, authUser } from "../../middleware/authSession";
+import { saltPassword } from "../../salter";
 
 // Gets all members
-router.get("/", async function (req, res) {
+router.get("/", authSession, async function (req, res) {
 	try {
-		db.User.find(function (err, users) {
-			if (err) return console.error(err);
-			res.json(users);
-		});
+		if (!res.headersSent) {
+			db.User.find(function (err, users) {
+				if (err) return console.error(err);
+				res.json(users);
+			});
+		}
 	} catch (err) {
 		console.error(err);
 		res.status(500).end();
@@ -41,30 +44,34 @@ router.get("/:username", authSession, async function (req, res) {
 router.post("/", async function (req, res) {
 	try {
 		let dateNow: number = Date.now();
-		// Hash here
+		if (
+			!req.body.username ||
+			!req.body.password ||
+			!req.body.displayName ||
+			!req.body.biography
+		) {
+			return res.status(400).send("Missing information");
+		}
+
+		let hash: any = saltPassword(req.body.password);
 
 		let newUser = new db.User({
 			username: req.body.username,
-			password: req.body.password,
+			hash: hash.hash,
+			salt: hash.salt,
 			displayName: req.body.displayName,
 			biography: req.body.biography,
 			creationDate: dateNow,
 			sessionKeys: [],
 		});
-		if (
-			!newUser.username ||
-			!newUser.password ||
-			!newUser.displayName ||
-			!newUser.biography
-		) {
-			return res.status(400).send("Missing information");
-		}
+
 		await newUser.save((err, newUser) => {
 			if (err) {
 				console.error(err);
 				return res.status(500).end();
 			}
 		});
+		//res.json(newUser.sendableUser());
 		res.json(newUser);
 	} catch (err) {
 		console.error(err);
@@ -73,18 +80,26 @@ router.post("/", async function (req, res) {
 });
 
 // Update member
-router.put("/:username", async function (req, res) {
+router.put("/:username", authSession, async function (req, res) {
 	try {
-		// Get requested post
-		// Update post if there is a change
+		if (!res.headersSent) {
+			// Get requested post
+			// Update post if there is a change
 
-		const userUpdate: any = {};
-		if (req.body.username) userUpdate.username = req.body.username;
-		if (req.body.password) userUpdate.password = req.body.password;
-		if (req.body.displayName) userUpdate.displayName = req.body.displayName;
-		if (req.body.biography) userUpdate.biography = req.body.biography;
-
-		db.User.updateOne({ username: req.params.username }, userUpdate);
+			const user: db.IUser = await db.User.findOne({
+				username: req.params.username,
+			});
+			if (user != null && user != undefined) {
+				if (authUser(req, res, user.id)) {
+					const userUpdate: any = {};
+					if (req.body.username) user.username = req.body.username;
+					if (req.body.displayName) user.displayName = req.body.displayName;
+					if (req.body.biography) user.biography = req.body.biography;
+					user.save();
+					res.json(user.sendableUser());
+				} else res.status(401).send("Not authorized to access this user");
+			} else res.status(404).end();
+		}
 	} catch (err) {
 		console.error(err);
 		res.status(500).end();
@@ -93,10 +108,12 @@ router.put("/:username", async function (req, res) {
 
 //Delete member
 
-router.delete("/:username", (req, res) => {
+router.delete("/:username", authSession, (req, res) => {
 	try {
-		db.User.deleteOne({ username: req.params.username });
-		res.end();
+		if (!res.headersSent) {
+			db.User.deleteOne({ username: req.params.username });
+			res.end();
+		}
 	} catch (err) {
 		console.error(err);
 		res.status(500).end();
